refactor time interpolation to helper module
