+++ workflows/prognostic_c48_run/runtime/nudging.py
@@ -2,28 +2,25 @@ import fv3gfs.util
 from mpi4py import MPI
 import shutil
 import fsspec
-import xarray as xr
 import dataclasses
 import cftime
 import functools
 from datetime import timedelta
 import os
 from typing import (
-    MutableMapping,
     Mapping,
     Iterable,
     Callable,
-    Hashable,
     Any,
     Dict,
     Optional,
 )
 import logging
+from .interpolate import time_interpolate_func
 from .names import STATE_NAME_TO_TENDENCY
+from .types import State
 
 logger = logging.getLogger(__name__)
-
-State = MutableMapping[Hashable, xr.DataArray]
 
 
 @dataclasses.dataclass
@@ -94,7 +91,7 @@ def setup_get_reference_state(
 
     initial_time_label = config.reference_initial_time
     if initial_time_label is not None:
-        get_reference_state = _time_interpolate_func(
+        get_reference_state = time_interpolate_func(
             get_reference_state,
             initial_time=_label_to_time(initial_time_label),
             frequency=timedelta(seconds=config.reference_frequency_seconds),
@@ -104,7 +101,7 @@ def setup_get_reference_state(
 
 
 def _get_reference_state(
-    time: str,
+    time: cftime.DatetimeJulian,
     reference_dir: str,
     communicator: fv3gfs.util.CubedSphereCommunicator,
     only_names: Iterable[str],
@@ -162,49 +159,6 @@ def _label_to_time(time: str) -> cftime.DatetimeJulian:
         int(time[11:13]),
         int(time[13:15]),
     )
-
-
-def _time_interpolate_func(
-    func: Callable[[cftime.DatetimeJulian], dict],
-    frequency: timedelta,
-    initial_time: cftime.DatetimeJulian,
-) -> Callable[[cftime.DatetimeJulian], dict]:
-    cached_func = functools.lru_cache(maxsize=2)(func)
-
-    @functools.wraps(cached_func)
-    def myfunc(time: cftime.DatetimeJulian) -> State:
-        quotient = (time - initial_time) // frequency
-        remainder = (time - initial_time) % frequency
-
-        state: State = {}
-        if remainder == timedelta(0):
-            state.update(cached_func(time))
-        else:
-            begin_time = quotient * frequency + initial_time
-            end_time = begin_time + frequency
-
-            state_0 = cached_func(begin_time)
-            state_1 = cached_func(end_time)
-
-            state.update(
-                _average_states(state_0, state_1, weight=(end_time - time) / frequency)
-            )
-
-        return state
-
-    return myfunc
-
-
-def _average_states(state_0: State, state_1: State, weight: float) -> State:
-    common_keys = set(state_0) & set(state_1)
-    out = {}
-    for key in common_keys:
-        if isinstance(state_1[key], xr.DataArray):
-            with xr.set_options(keep_attrs=True):
-                out[key] = (
-                    state_0[key] * weight + (1 - weight) * state_1[key]  # type: ignore
-                )
-    return out
 
 
 def get_nudging_tendency(
